Elastic-ViT-Input-Adaptive/elastic_vit/config.py
from __future__ import annotations

from copy import deepcopy
from pathlib import Path
from typing import Any, Dict

import yaml


def _deep_update(base: Dict[str, Any], override: Dict[str, Any]) -> Dict[str, Any]:
    merged = deepcopy(base)
    for key, value in override.items():
        if isinstance(value, dict) and isinstance(merged.get(key), dict):
            merged[key] = _deep_update(merged[key], value)
        else:
            merged[key] = value
    return merged


def load_config(path: str | Path) -> Dict[str, Any]:
    config_path = Path(path)
    with config_path.open("r", encoding="utf-8") as handle:
        config = yaml.safe_load(handle)

    if not isinstance(config, dict):
        raise ValueError(f"Expected a mapping in config file: {config_path}")

    base_path = config.pop("base_config", None)
    if base_path:
        parent = config_path.parent / base_path
        base = load_config(parent)
        config = _deep_update(base, config)

    config["config_path"] = str(config_path.resolve())
    return config

