Elastic-ViT-Input-Adaptive/requirements.txt
torch>=2.2
torchvision>=0.17
timm>=1.0
PyYAML>=6.0
tqdm>=4.66
pycocotools>=2.0
